build(CLI11): Add argument parsing library
Looked into quite a few of these libraries. I mainly just wanted
something with nice help functionality (separation of positional and
optional arguments) and something that doesn't clutter the main function
(boost). There was also the option of using 'argparse', but CLI11 has
way more documentation available. It is probably way more than I need,
but others attest to it, so we'll stick it out.