perf: add binary search algorithm for cpg indexes
This is will be faster than the two pointer method if the number of CpGs
in the cpg_list is less than (bedmethyl_rows/log2(bedmethyl_rows)).
Assuming normal bedmethyl files, with all CpGs included, this magic
number is around 100,000 CpGs. (So if the list is smaller than this, we
should go with the binary search method, otherwise we should use two
pointer approach)