fix: add numerical regularisation to avoid negative eigenvalues
Negative eigenvalues cause the LLT cholesky decomposition (required as a
part of the Goldfarb&Idnani solver) to fail as this implies that the
matrix is not positive definite. A regularisation term is applied to the
main diagonal (adding 1e-8 to each value) in an attempt to resolve this.
A further check might be required in the future to increase this value
if 1e-8 fails (though this is probably unlikely).