perf: remove need to parse end position
cpg sites span a single base and so there 'should' only be two bits of
data for each start position (one for m, one for h). I say should,
because technically any bed file with >=11 fields can be given here. But
that's on the user (and wasn't being accounted for before either).
Checking for duplicate rows in terms of keys would be time consuming.