docs: Correctly derive the QPP
Damn, I had this wrong for a long time and was always wondering about
the factor of a half that went missing. This is now fixed, glad we
figured that one out eventually (the maths was fine, but the start
objective function was not, which is probably why I didn't catch it).