refactor: modernise file reading (#3)
* perf: Only add warning messages to vector if they will be shown

There is no need to store 1000s of warning messages if only 5 are going
to be non-supressed. This just causes a bunch of reallocations and
unecessary insertions.

* style: use brace initialisation over copy

* refactor: use std::filesystem over strings for paths

* refactor: use emplace_back over push_back in a couple of scenarios

This reduces initialisation followed by moves (but only in C++23)

* feat: add a RAII compliant file descriptor class

To be used later by TSV reader class

* feat: add a RAII compliant memory map class

* refactor: implement RAII compliant classes into TSVReader class

* refactor: Use a map range pair over getting start/end over and over

* refactor: make this constant constexpr

* refactor: replace pair with formal struct

* chore: remove unneeded header

* chore: update comment to reflect load

* refactor: use teardown over reimplementing logic

* refactor: Use consistent exception types in FileDescriptor

* test: update tests to work with new exception logic

Exceptions are no longer system errors and some occur on initialisation
instead of on load (file descriptor and memory mapping based errors)