feat: update convergence criteria to use l2 norm of objective function
The change in cell proporitons might not change a whole lot between
iterations, but the reference matrix does. It is better to break when
the l2 norm of the objective function (l2 norm of bulk.T -
reference*cell_proportions) is very small. In such an event, then our
particular set of proportions is very close to an actual solution.